refactor name ReceptViewModel to CreateReceptViewModel.kt create ReceptViewModel.kt for read only recept screen add navigation arg for open recept screen by recept id
