+++ .idea/misc.xml
@@ -1,25 +1,6 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <project version="4">
-  <component name="DesignSurface">
-    <option name="filePathToZoomLevelMap">
-      <map>
-        <entry key="..\:/Users/j63ch/AndroidStudioProjects/Confectioners_organizer/app/src/main/res/drawable-anydpi/ic_action_confirm.xml" value="0.14629629629629629" />
-        <entry key="..\:/Users/j63ch/AndroidStudioProjects/Confectioners_organizer/app/src/main/res/drawable/ic_baseline_add_circle_24.xml" value="0.13240740740740742" />
-        <entry key="..\:/Users/j63ch/AndroidStudioProjects/Confectioners_organizer/app/src/main/res/drawable/ic_baseline_add_circle_outline_24.xml" value="0.13240740740740742" />
-        <entry key="..\:/Users/j63ch/AndroidStudioProjects/Confectioners_organizer/app/src/main/res/drawable/ic_baseline_arrow_back_24.xml" value="0.14629629629629629" />
-        <entry key="..\:/Users/j63ch/AndroidStudioProjects/Confectioners_organizer/app/src/main/res/drawable/ic_baseline_check_circle_outline_24.xml" value="0.14629629629629629" />
-        <entry key="..\:/Users/j63ch/AndroidStudioProjects/Confectioners_organizer/app/src/main/res/drawable/ic_baseline_content_paste_24.xml" value="0.12777777777777777" />
-        <entry key="..\:/Users/j63ch/AndroidStudioProjects/Confectioners_organizer/app/src/main/res/drawable/ic_baseline_dehaze_24.xml" value="0.13055555555555556" />
-        <entry key="..\:/Users/j63ch/AndroidStudioProjects/Confectioners_organizer/app/src/main/res/drawable/ic_baseline_delivery_dining_24.xml" value="0.13055555555555556" />
-        <entry key="..\:/Users/j63ch/AndroidStudioProjects/Confectioners_organizer/app/src/main/res/drawable/ic_baseline_edit_24.xml" value="0.13240740740740742" />
-        <entry key="..\:/Users/j63ch/AndroidStudioProjects/Confectioners_organizer/app/src/main/res/drawable/ic_baseline_location_on_24.xml" value="0.13055555555555556" />
-        <entry key="..\:/Users/j63ch/AndroidStudioProjects/Confectioners_organizer/app/src/main/res/drawable/ic_baseline_notifications_active_24.xml" value="0.13240740740740742" />
-        <entry key="..\:/Users/j63ch/AndroidStudioProjects/Confectioners_organizer/app/src/main/res/drawable/ic_launcher_background.xml" value="0.13240740740740742" />
-        <entry key="..\:/Users/j63ch/AndroidStudioProjects/Confectioners_organizer/app/src/main/res/drawable/ic_round_assignment_24.xml" value="0.12777777777777777" />
-      </map>
-    </option>
-  </component>
-  <component name="ProjectRootManager" version="2" languageLevel="JDK_11" default="true" project-jdk-name="Android Studio default JDK" project-jdk-type="JavaSDK">
+  <component name="ProjectRootManager" version="2" languageLevel="JDK_11" default="true" project-jdk-name="1.8" project-jdk-type="JavaSDK">
     <output url="file://$PROJECT_DIR$/build/classes" />
   </component>
   <component name="ProjectType">
